Log expander steps and linklet notes to error port
This uses `eprintf` (from Racket) in the expander with the hope that any traced
passes (which use Chez printing APIs) print in the expected ordering relative to
expander steps and linklet notes. (If you send expander / linklet-layer and
Chez-layer output to the same port via different printing APIs, you seem to lose
the all-important time ordering property.)

The error port is specifically used so that logs won't break programs expecting
particular output on the output port (such as `schemify` steps when developing
the expander itself).

The assumptions here may now be invalid, since the previous commit changes the
Chez pass tracer to use the error port as well. It may instead be necessary to
have all layers use the Chez printing APIs...