Log expander steps and linklet notes using Chez printing APIs
Ideally we would just use the standard Racket logging facility.

Unfortunately we are faced with a somewhat unusual requirement to produce a
consistently ordered event trace that weaves together messages from several
different layers:

  - `cs/linklet.sls`
  - `expander/**`
  - `ChezScheme/s/**`

To achieve this, we reach for the lowest layer (Chez) and expose its functions
as needed in the other layers.

This uses the same output path as Chez's nanopass tracer, which helps ensure
trace logs from all layers appear together in the order they occurred.